move integration tests back
