+++ launcher/src/backend/ethereum-services/PrysmBeaconService.int.js
@@ -1,11 +1,11 @@
 /**
  * @jest-environment node
  */
-import { HetznerServer } from "../../HetznerServer.js";
-import { NodeConnection } from "../../NodeConnection.js";
-import { ServiceManager } from "../../ServiceManager.js";
-import { TaskManager } from "../../TaskManager.js";
-import { ValidatorAccountManager } from "../../ValidatorAccountManager.js";
+import { HetznerServer } from "../HetznerServer.js";
+import { NodeConnection } from "../NodeConnection.js";
+import { ServiceManager } from "../ServiceManager.js";
+import { TaskManager } from "../TaskManager.js";
+import { ValidatorAccountManager } from "../ValidatorAccountManager.js";
 const log = require("electron-log");
 
 jest.setTimeout(1000000);
